Updated the Interface
This release is a lot faster in updating the questions for each student.
I had to rewrite some stuff in a few files for this, but it's pretty
snazzy now.  You now have to pick the directory you'd like to load files
from, and also you need to pick which dictionary you're going to be
using.  For the first lab we're looking at using lab1.dat.  This is
coming together nicely.