Some Quick Bug Fixes
I added some quick and dirty error handling for when we have missing
information.  One of our questions (#9) had the wrong aText so it wasn't
finding the answer properly.  It was working okay on our text examples,
but not on my students assignment.  I also made it so it will strip
leading 0's from section numbers because Section 09 and Section 9 were
being differentiated.  I also had a student who didn't put his name or
section so I made a quick catch for that.  I also strip the students
answer for leading and trailing spaces because it'll make things easier.