Added basic support for equations in word documents.
I added a really simplistic way of getting math equation objects from
the word documents.  I may try a different approach later.  For now the
items aren't being shown, but if you want you can look in the code and
turn the box on to see what kind of objects we've got.  I also added a
little bit of error handling and reporting to the data extraction.