Ability to extract images from documents too, but turned off for now.
We can now extract the images saved in the word documents too, but it
has to be saved to the folder it's currently in and we have to use the
clipboard to do it.  It's not the best way to do things, but it at least
works.  I've got it disabled for now because it slows things down, but
you can turn it on if you'd like to try it.