Finished Class Rewrite
I finished rewriting the stuff into classes and fixed a few things along
the way.  It should look and act the same as before.  But now we can
call and access anything we need.  We can also build out wrappers or
helpers for things we want to do in certain areas if they're happening a
lot.