HUGE Rewrite, Read Description
I was getting frustrated about having 3 different databases with similar
information.  I didn't realize we already had some info in the
question_bank that help points and weight, etc.  So I threw it all out
and started over fresh with one master database to hold the assignments,
comments, questionbank etc.  It's a lot easier to work with I think and
I'm gonna work on a way to save the database so you can do partial work.
The only thing that I don't think is currently working right now is the
automated grading.  I'll get it up and running soon though.