Moved Cheat Check
It was in the wrong function so it wasn't being called when the students
information was actually in the tree.  It didn't find the two students
who copied eachother though.  Possible because of different spacing.
Maybe we should strip all new lines and spaces and just check how
similar the text is?