Excel Integration
Hopefully this doesn't break anything.  I've included an update to
integrate excel so we can grade those automatically.  It's pretty basic
right now but I had to rename some stuff to make sure that we can
distinguish between word files and excel files when getting files etc.
But yeah, this is pretty awesome.