Comments Update
Made the Comments a lot better.  I changed it so that it doesn't put
what their answer was or what the correct answer was.  Instead now I've
added a new "reason" field so that we can write out a detailed
explanation that'll get auto put in there so that they can see it now
instead.  It also shows their grade for that particular question so they
can see how much they actually got for it.  Also the comments now
automatically populate and update if you remove or change the grading so
that you don't have to refresh it.  The downside is that you'll lose any
custom comments but there really shouldn't be as many of those now.