Finished Lab 4, Added Unicode Back
I finished the auto grading portion of stuff for Lab #4.  I also added
unicode support back.  It works for me, but you'll have to let me know.
I had to use unicode because some people were using dash symbols instead
of - so that it was marking things wrong when it shouldn't.