Quick Opening & Saving Fixes
Just added a way for the import wizard to auto filter specific items so
that it is easier to see things.  I name my attendance sheets so I'd
like to filter based on them.  I made it so you can filter based on all
csv or all files if needed though.