Pity Points & More Lenient Grading
I added a button to give a quick 1 point (or half a point if the
question is worth only 1 point) so that we can give students points for
at least trying pretty easily.  I also added a more lenient
autoweighting system so that it will incorporate half point steps when
appropriate.  Additionally I added the points that the student has for
this question above it so we can see what they've got.